+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 -   Commands:
     -   Ds
     -   Sc
+-   Clipboard class.
 
 ### Changed